requirements.txt
bcrypt==3.2.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
flask-mongoengine
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.20.0
Pillow==8.2.0
pymongo==3.11.0
python-dotenv==0.14.0
requests==2.24.0
Werkzeug==1.0.1
WTForms==2.3.3